Breaking different implementations into different headers
