+++ src/Concurrency/tests/test_TaskBlock.cpp
@@ -16,7 +16,7 @@
 #define CATCH_CONFIG_MAIN
 #include "catch.hpp"
 
-#include "Concurrency/TaskBlock.hpp"
+#include "Concurrency/TasksOneToOne.hpp"
 
 namespace qmcplusplus
 {
